+++ src/chart/types/traffic_chart.rs
@@ -9,12 +9,13 @@ use plotters::prelude::*;
 use plotters_iced::{Chart, ChartBuilder, ChartWidget, DrawingBackend};
 use splines::Spline;
 
-use crate::gui::app::FONT_FAMILY_NAME;
+use crate::gui::sniffer::FONT_FAMILY_NAME;
 use crate::gui::styles::style_constants::CHARTS_LINE_BORDER;
 use crate::gui::styles::types::palette::to_rgb_color;
 use crate::gui::types::message::Message;
 use crate::networking::types::traffic_direction::TrafficDirection;
 use crate::translations::translations::{incoming_translation, outgoing_translation};
+use crate::utils::formatted_strings::get_formatted_num_seconds;
 use crate::{ByteMultiple, ChartType, Language, StyleType};
 
 /// Struct defining the chart to be displayed in gui run page
@@ -118,7 +119,7 @@ impl TrafficChart {
         min - gap..max + gap
     }
 
-    fn font(&self, size: f64) -> TextStyle<'static> {
+    fn font<'a>(&self, size: f64) -> TextStyle<'a> {
         (FONT_FAMILY_NAME, size)
             .into_font()
             .style(self.style.get_font_weight())
@@ -217,7 +218,10 @@ impl Chart<Message> for TrafficChart {
                 &|bytes| ByteMultiple::formatted_string(bytes.abs() as u128)
             })
             .x_labels(min(6, x_labels))
-            .x_label_formatter(&std::string::ToString::to_string)
+            .x_label_formatter(
+                #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+                &|seconds| get_formatted_num_seconds(seconds.abs() as u128),
+            )
             .draw()
             .unwrap();