+++ src/configs/types/configs.rs
@@ -1,4 +1,7 @@
 use crate::{ConfigDevice, ConfigSettings, ConfigWindow};
+use once_cell::sync::Lazy;
+
+pub static CONFIGS: Lazy<Configs> = Lazy::new(Configs::load);
 
 #[derive(Default, Clone, PartialEq, Debug)]
 pub struct Configs {
@@ -8,6 +11,8 @@ pub struct Configs {
 }
 
 impl Configs {
+    /// This should only be used directly to load fresh configs;
+    /// use `CONFIGS` instead to access the initial instance
     pub fn load() -> Self {
         Configs {
             settings: ConfigSettings::load(),