+++ src/countries/country_utils.rs
@@ -28,14 +28,14 @@ use crate::translations::translations_2::{
 };
 use crate::{Language, StyleType};
 
-fn get_flag_from_country(
+fn get_flag_from_country<'a>(
     country: Country,
     width: f32,
     is_local: bool,
     is_loopback: bool,
     traffic_type: TrafficType,
     language: Language,
-) -> (Svg<StyleType>, String) {
+) -> (Svg<'a, StyleType>, String) {
     #![allow(clippy::too_many_lines)]
     let mut tooltip = country.to_string();
     let mut svg_style = SvgType::Standard;
@@ -303,20 +303,20 @@ fn get_flag_from_country(
             flag
         }
     })))
-    .style(svg_style)
+    .class(svg_style)
     .width(width)
     .height(width * 0.75);
 
     (svg, tooltip)
 }
 
-pub fn get_flag_tooltip(
+pub fn get_flag_tooltip<'a>(
     country: Country,
     host_info: &DataInfoHost,
     language: Language,
     font: Font,
     thumbnail: bool,
-) -> Tooltip<'static, Message, StyleType> {
+) -> Tooltip<'a, Message, StyleType> {
     let width = if thumbnail {
         FLAGS_WIDTH_SMALL
     } else {
@@ -346,7 +346,7 @@ pub fn get_flag_tooltip(
         Position::FollowCursor,
     )
     .snap_within_viewport(true)
-    .style(tooltip_style);
+    .class(tooltip_style);
 
     if width == FLAGS_WIDTH_SMALL {
         tooltip = tooltip.padding(3);
@@ -355,13 +355,13 @@ pub fn get_flag_tooltip(
     tooltip
 }
 
-pub fn get_computer_tooltip(
+pub fn get_computer_tooltip<'a>(
     is_my_address: bool,
     is_local: bool,
     traffic_type: TrafficType,
     language: Language,
     font: Font,
-) -> Tooltip<'static, Message, StyleType> {
+) -> Tooltip<'a, Message, StyleType> {
     let content = Svg::new(Handle::from_memory(Vec::from(
         match (is_my_address, is_local, traffic_type) {
             (true, _, _) => COMPUTER,
@@ -371,7 +371,7 @@ pub fn get_computer_tooltip(
             (false, false, TrafficType::Unicast) => UNKNOWN,
         },
     )))
-    .style(SvgType::AdaptColor)
+    .class(SvgType::AdaptColor)
     .width(FLAGS_WIDTH_BIG)
     .height(FLAGS_WIDTH_BIG * 0.75);
 
@@ -390,5 +390,5 @@ pub fn get_computer_tooltip(
         Position::FollowCursor,
     )
     .snap_within_viewport(true)
-    .style(ContainerType::Tooltip)
+    .class(ContainerType::Tooltip)
 }