+++ src/gui/mod.rs
@@ -1,5 +1,5 @@
-pub mod app;
 pub mod components;
 pub mod pages;
+pub mod sniffer;
 pub mod styles;
 pub mod types;