+++ src/gui/pages/settings_notifications_page.rs
@@ -1,8 +1,7 @@
-use iced::alignment::{Horizontal, Vertical};
 use iced::widget::scrollable::Direction;
 use iced::widget::{horizontal_space, Button, Slider};
 use iced::widget::{Checkbox, Column, Container, Row, Scrollable, Space, Text, TextInput};
-use iced::{Alignment, Font, Length};
+use iced::{Alignment, Font, Length, Padding};
 
 use crate::gui::components::button::button_hide;
 use crate::gui::components::tab::get_settings_tabs;
@@ -27,7 +26,7 @@ use crate::translations::translations::{
 use crate::utils::types::icon::Icon;
 use crate::{ConfigSettings, Language, Sniffer, StyleType};
 
-pub fn settings_notifications_page(sniffer: &Sniffer) -> Container<Message, StyleType> {
+pub fn settings_notifications_page<'a>(sniffer: &Sniffer) -> Container<'a, Message, StyleType> {
     let ConfigSettings {
         style,
         language,
@@ -55,54 +54,52 @@ pub fn settings_notifications_page(sniffer: &Sniffer) -> Container<Message, Styl
         .push(
             notifications_title_translation(language)
                 .font(font)
-                .style(TextType::Subtitle)
+                .class(TextType::Subtitle)
                 .size(FONT_SIZE_SUBTITLE)
                 .width(Length::Fill)
-                .horizontal_alignment(Horizontal::Center),
+                .align_x(Alignment::Center),
         )
         .push(Space::with_height(5));
 
     let volume_notification_col = Column::new()
-        .padding([0, 0, 5, 0])
-        .align_items(Alignment::Center)
+        .padding(Padding::ZERO.bottom(5))
+        .align_x(Alignment::Center)
         .width(Length::Fill)
         .push(volume_slider(language, font, notifications.volume))
-        .push(
-            Scrollable::new(
-                Column::new()
-                    .width(720)
-                    .push(get_packets_notify(
-                        notifications.packets_notification,
-                        language,
-                        font,
-                    ))
-                    .push(get_bytes_notify(
-                        notifications.bytes_notification,
-                        language,
-                        font,
-                    ))
-                    .push(get_favorite_notify(
-                        notifications.favorite_notification,
-                        language,
-                        font,
-                    )),
-            )
-            .direction(Direction::Vertical(ScrollbarType::properties())),
-        );
+        .push(Scrollable::with_direction(
+            Column::new()
+                .width(720)
+                .push(get_packets_notify(
+                    notifications.packets_notification,
+                    language,
+                    font,
+                ))
+                .push(get_bytes_notify(
+                    notifications.bytes_notification,
+                    language,
+                    font,
+                ))
+                .push(get_favorite_notify(
+                    notifications.favorite_notification,
+                    language,
+                    font,
+                )),
+            Direction::Vertical(ScrollbarType::properties()),
+        ));
 
     content = content.push(volume_notification_col);
 
     Container::new(content)
         .height(400)
         .width(800)
-        .style(ContainerType::Modal)
+        .class(ContainerType::Modal)
 }
 
-fn get_packets_notify(
+fn get_packets_notify<'a>(
     packets_notification: PacketsNotification,
     language: Language,
     font: Font,
-) -> Column<'static, Message, StyleType> {
+) -> Column<'a, Message, StyleType> {
     let checkbox = Checkbox::new(
         packets_threshold_translation(language),
         packets_notification.threshold.is_some(),
@@ -136,7 +133,7 @@ fn get_packets_notify(
             Container::new(ret_val)
                 .padding(10)
                 .width(700)
-                .style(ContainerType::BorderedRound),
+                .class(ContainerType::BorderedRound),
         )
     } else {
         let input_row = input_group_packets(packets_notification, font, language);
@@ -146,16 +143,16 @@ fn get_packets_notify(
             Container::new(ret_val)
                 .padding(10)
                 .width(700)
-                .style(ContainerType::BorderedRound),
+                .class(ContainerType::BorderedRound),
         )
     }
 }
 
-fn get_bytes_notify(
+fn get_bytes_notify<'a>(
     bytes_notification: BytesNotification,
     language: Language,
     font: Font,
-) -> Column<'static, Message, StyleType> {
+) -> Column<'a, Message, StyleType> {
     let checkbox = Checkbox::new(
         bytes_threshold_translation(language),
         bytes_notification.threshold.is_some(),
@@ -189,7 +186,7 @@ fn get_bytes_notify(
             Container::new(ret_val)
                 .padding(10)
                 .width(700)
-                .style(ContainerType::BorderedRound),
+                .class(ContainerType::BorderedRound),
         )
     } else {
         let input_row = input_group_bytes(bytes_notification, font, language);
@@ -199,16 +196,16 @@ fn get_bytes_notify(
             Container::new(ret_val)
                 .padding(10)
                 .width(700)
-                .style(ContainerType::BorderedRound),
+                .class(ContainerType::BorderedRound),
         )
     }
 }
 
-fn get_favorite_notify(
+fn get_favorite_notify<'a>(
     favorite_notification: FavoriteNotification,
     language: Language,
     font: Font,
-) -> Column<'static, Message, StyleType> {
+) -> Column<'a, Message, StyleType> {
     let checkbox = Checkbox::new(
         favorite_notification_translation(language),
         favorite_notification.notify_on_favorite,
@@ -239,26 +236,26 @@ fn get_favorite_notify(
             Container::new(ret_val)
                 .padding(10)
                 .width(700)
-                .style(ContainerType::BorderedRound),
+                .class(ContainerType::BorderedRound),
         )
     } else {
         Column::new().padding(5).push(
             Container::new(ret_val)
                 .padding(10)
                 .width(700)
-                .style(ContainerType::BorderedRound),
+                .class(ContainerType::BorderedRound),
         )
     }
 }
 
-fn input_group_packets(
+fn input_group_packets<'a>(
     packets_notification: PacketsNotification,
     font: Font,
     language: Language,
-) -> Container<'static, Message, StyleType> {
+) -> Container<'a, Message, StyleType> {
     let curr_threshold_str = &packets_notification.threshold.unwrap().to_string();
     let input_row = Row::new()
-        .align_items(Alignment::Center)
+        .align_y(Alignment::Center)
         .spacing(5)
         .push(Space::with_width(45))
         .push(Text::new(format!("{}:", threshold_translation(language))).font(font))
@@ -286,19 +283,19 @@ fn input_group_packets(
         .push(
             Text::new(per_second_translation(language))
                 .font(font)
-                .vertical_alignment(Vertical::Center)
+                .align_y(Alignment::Center)
                 .size(FONT_SIZE_FOOTER),
         );
     Container::new(input_row)
-        .align_x(Horizontal::Center)
-        .align_y(Vertical::Center)
+        .align_x(Alignment::Center)
+        .align_y(Alignment::Center)
 }
 
-fn input_group_bytes(
+fn input_group_bytes<'a>(
     bytes_notification: BytesNotification,
     font: Font,
     language: Language,
-) -> Container<'static, Message, StyleType> {
+) -> Container<'a, Message, StyleType> {
     let info_str = format!(
         "{}; {}",
         per_second_translation(language),
@@ -310,7 +307,7 @@ fn input_group_bytes(
     curr_threshold_str.push_str(&bytes_notification.byte_multiple.get_char());
     let input_row = Row::new()
         .spacing(5)
-        .align_items(Alignment::Center)
+        .align_y(Alignment::Center)
         .push(Space::with_width(45))
         .push(Text::new(format!("{}:", threshold_translation(language))).font(font))
         .push(
@@ -333,23 +330,23 @@ fn input_group_bytes(
         .push(
             Text::new(info_str)
                 .font(font)
-                .vertical_alignment(Vertical::Center)
+                .align_y(Alignment::Center)
                 .size(FONT_SIZE_FOOTER),
         );
     Container::new(input_row)
-        .align_x(Horizontal::Center)
-        .align_y(Vertical::Center)
+        .align_x(Alignment::Center)
+        .align_y(Alignment::Center)
 }
 
-fn volume_slider(
+fn volume_slider<'a>(
     language: Language,
     font: Font,
     volume: u8,
-) -> Container<'static, Message, StyleType> {
+) -> Container<'a, Message, StyleType> {
     Container::new(
         Column::new()
             .spacing(5)
-            .align_items(Alignment::Center)
+            .align_x(Alignment::Center)
             .push(Text::new(format!("{}: {volume:^3}%", volume_translation(language))).font(font))
             .push(
                 Row::new()
@@ -357,7 +354,7 @@ fn volume_slider(
                         Icon::AudioMute
                             .to_text()
                             .width(30)
-                            .vertical_alignment(Vertical::Center)
+                            .align_y(Alignment::Center)
                             .size(20),
                     )
                     .push(
@@ -369,7 +366,7 @@ fn volume_slider(
                     .push(
                         Icon::AudioHigh
                             .to_text()
-                            .vertical_alignment(Vertical::Center)
+                            .align_y(Alignment::Center)
                             .size(20),
                     ),
             ),
@@ -377,15 +374,15 @@ fn volume_slider(
     .padding(5)
     .width(Length::Fill)
     .height(60)
-    .align_x(Horizontal::Center)
-    .align_y(Vertical::Center)
+    .align_x(Alignment::Center)
+    .align_y(Alignment::Center)
 }
 
-fn sound_buttons(
+fn sound_buttons<'a>(
     notification: Notification,
     font: Font,
     language: Language,
-) -> Row<'static, Message, StyleType> {
+) -> Row<'a, Message, StyleType> {
     let current_sound = match notification {
         Notification::Packets(n) => n.sound,
         Notification::Bytes(n) => n.sound,
@@ -393,7 +390,7 @@ fn sound_buttons(
     };
 
     let mut ret_val = Row::new()
-        .align_items(Alignment::Center)
+        .align_y(Alignment::Center)
         .spacing(5)
         .push(Space::with_width(45))
         .push(Text::new(format!("{}:", sound_translation(language))).font(font));
@@ -414,7 +411,7 @@ fn sound_buttons(
                 .padding(0)
                 .width(80)
                 .height(25)
-                .style(if is_active {
+                .class(if is_active {
                     ButtonType::BorderedRoundSelected
                 } else {
                     ButtonType::BorderedRound
@@ -428,12 +425,12 @@ fn sound_buttons(
     ret_val
 }
 
-pub fn settings_header(
+pub fn settings_header<'a>(
     font: Font,
     font_headers: Font,
     color_gradient: GradientType,
     language: Language,
-) -> Container<'static, Message, StyleType> {
+) -> Container<'a, Message, StyleType> {
     Container::new(
         Row::new()
             .push(horizontal_space())
@@ -442,17 +439,17 @@ pub fn settings_header(
                     .font(font_headers)
                     .size(FONT_SIZE_TITLE)
                     .width(Length::FillPortion(6))
-                    .horizontal_alignment(Horizontal::Center),
+                    .align_x(Alignment::Center),
             )
             .push(
                 Container::new(button_hide(Message::CloseSettings, language, font))
                     .width(Length::Fill)
-                    .align_x(Horizontal::Center),
+                    .align_x(Alignment::Center),
             ),
     )
-    .align_x(Horizontal::Center)
-    .align_y(Vertical::Center)
+    .align_x(Alignment::Center)
+    .align_y(Alignment::Center)
     .height(40)
     .width(Length::Fill)
-    .style(ContainerType::Gradient(color_gradient))
+    .class(ContainerType::Gradient(color_gradient))
 }